requirements.txt
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.2
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
